+++ src/function_highlighting/decorations.test.ts
@@ -28,4 +28,3 @@ suite("Function highlighting decorations", () => {
 		);
 	});
 });
-